feat: add unit tests for course class controller
This commit introduces unit tests for the `CourseClassController` class, covering the following functionalities:

- Deleting a course class by ID.
- Updating a course class by ID.
- Creating a new course class.
- Retrieving students enrolled in a course class.
- Removing a student from a course class.
- Adding a student to a course class.

The tests ensure the proper functioning of the controller methods and their interactions with the `Repository`. Some tests are currently commented out due to errors encountered when running outside of a module. These tests are marked with "EXPECTED ERROR" and "TODO" comments for future investigation and refactoring.