feat: add FLASK_APP and FLASK_ENV to .env.example
This commit adds `FLASK_APP` and `FLASK_ENV` variables to the `.env.example` file. These variables are used to configure the Flask application, specifying the entry point and the environment in which the application is running. This provides a more complete example for users setting up the application for development.